+++ multiqc/modules/salmon/salmon.py
@@ -1,0 +1,93 @@
+#!/usr/bin/env python
+
+""" MultiQC module to parse output from Salmon """
+
+from __future__ import print_function
+from collections import OrderedDict
+import json
+import logging
+import os
+
+from multiqc import config, BaseMultiqcModule, plots
+
+# Initialise the logger
+log = logging.getLogger(__name__)
+
+class MultiqcModule(BaseMultiqcModule):
+
+    def __init__(self):
+
+        # Initialise the parent object
+        super(MultiqcModule, self).__init__(name='Salmon', anchor='salmon',
+        href='http://combine-lab.github.io/salmon/', 
+        info="is a tool for quantifying the expression of transcripts using RNA-seq data.")
+        
+        # Parse meta information. JSON win!
+        self.salmon_meta = dict()
+        for f in self.find_log_files(config.sp['salmon']['meta']):
+            # Get the s_name from the parent directory
+            if os.path.basename(f['root']) == 'aux':
+                s_name = os.path.basename( os.path.dirname(f['root']) )
+                self.salmon_meta[s_name] = json.loads(f['f'])
+        
+        # Parse Fragment Length Distribution logs
+        self.salmon_fld = dict()
+        for f in self.find_log_files(config.sp['salmon']['fld']):
+            # Get the s_name from the parent directory
+            if os.path.basename(f['root']) == 'libParams':
+                s_name = os.path.basename( os.path.dirname(f['root']) )
+                parsed = OrderedDict()
+                for i, v in enumerate(f['f'].split()):
+                    parsed[i] = float(v)
+                if len(parsed) > 0:
+                    if s_name in self.salmon_fld:
+                        log.debug("Duplicate sample name found! Overwriting: {}".format(s_name))
+                    self.add_data_source(f, s_name)
+                    self.salmon_fld[s_name] = parsed
+
+        if len(self.salmon_meta) == 0 and len(self.salmon_fld) == 0:
+            log.debug("Could not find any Salmon data in {}".format(config.analysis_dir))
+            raise UserWarning
+
+        if len(self.salmon_meta) > 0:
+            log.info("Found {} meta reports".format(len(self.salmon_meta)))
+            self.write_data_file(self.salmon_meta, 'multiqc_salmon')
+        if len(self.salmon_fld) > 0:
+            log.info("Found {} fragment length distributions".format(len(self.salmon_fld)))
+        
+        # Add alignment rate to the general stats table
+        headers = OrderedDict()
+        headers['percent_mapped'] = {
+            'title': '% Aligned',
+            'description': '% Mapped reads',
+            'max': 100,
+            'min': 0,
+            'suffix': '%',
+            'scale': 'YlGn',
+            'format': '{:.1f}%'
+        }
+        headers['num_mapped'] = {
+            'title': 'M Aligned',
+            'description': 'Mapped reads (millions)',
+            'min': 0,
+            'scale': 'PuRd',
+            'modify': lambda x: float(x) / 1000000,
+            'shared_key': 'read_count'
+        }
+        self.general_stats_addcols(self.salmon_meta, headers)
+        
+        # Fragment length distribution plot
+        # Only one section, so add to the intro
+        pconfig = {
+            'smooth_points': 500,
+            'id': 'salmon_plot',
+            'title': 'Salmon Fragment Length Distribution',
+            'ylab': 'Fraction',
+            'xlab': 'Fragment Length (bp)',
+            'ymin': 0,
+            'xmin': 0,
+            'tt_label': '<b>{point.x:,.0f} bp</b>: {point.y:,.0f}',
+        }
+        
+        self.intro += plots.linegraph.plot(self.salmon_fld, pconfig)
+